ModelDynamics

ModelDynamics

ModelDynamics

Details

Represents a reusable learning strategy composed of a partner selection function, an interaction function, and a descriptive label.

Methods

Public methods

ModelDynamics$new()

ModelDynamics$get_partner_selection()

ModelDynamics$get_interaction()

ModelDynamics$get_model_step()

ModelDynamics$get_label()

ModelDynamics$clone()

Method new()

Create a new ModelDynamics

Usage

ModelDynamics$new(partner_selection, interaction, model_step, label)

Arguments

partner_selection

Function to select the teacher (or NULL)

interaction

Function to define the interaction (or NULL)

label

A string label for this strategy

Method get_partner_selection()

Get the partner selection function

Usage

ModelDynamics$get_partner_selection()

Returns

Function

Method get_interaction()

Get the interaction function

Usage

ModelDynamics$get_interaction()

Returns

Function

Method get_model_step()

Get the interaction function

Usage

ModelDynamics$get_model_step()

Returns

Function

Method get_label()

Get the strategy label

Usage

ModelDynamics$get_label()

Returns

Character string

Method clone()

The objects of this class are cloneable with this method.

Usage

ModelDynamics$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.