Run a grid of trial ensembles with parameter metadata

Runs trial ensembles across a parameter grid. All scalar and function-valued parameters used in model construction or trial dynamics are included in metadata for transparency.

Usage

run_trials(
  model_generator,
  n_trials_per_param = 10,
  stop = 10,
  .progress = TRUE,
  observe = "behavior",
  syncfile = NULL,
  overwrite = FALSE,
  ...
)

Arguments

model_generator

Function that returns a new AgentBasedModel instance according to model_parameters, a named list of parameter label-value pairs.

n_trials_per_param

Number of trials per parameter combination.

stop

Stopping condition (number or function).

.progress

Whether to show progressbar when running the trials.

...

List of parameter label-value pairs; vector or singleton values.

Value

A list of Trial objects

Examples

abm_gen <- function(params) {
  params$graph <- make_small_world(params$n_agents, 6, 0.5)
  return (do.call(make_abm, params) %>%
            initialize_agents(
              initial_prevalence = params$initial_prevalence,
              adaptive_fitness = params$adaptive_fitness
            )
  )
}
trials <-
  run_trials(
    abm_gen,
    n_trials_per_param = 2,
    stop = socmod::fixated,
    n_agents = 20,
    initial_prevalence = 0.1,
    adaptive_fitness = c(0.9, 1.1, 1.3)
)